feat: move types folder to out of app folder
