+++ src/app/actions/users/update-user-action.ts
@@ -3,10 +3,9 @@
 import { redirect } from 'next/navigation'
 import { revalidatePath } from 'next/cache'
 
+import { User } from '@/types'
 import { updateUser } from '@/app/services/api-users'
 import { getErrorMessage } from '@/app/utils/getErrorMessage'
-
-import { User } from '@/app/types'
 
 export const updateUserServerAction = async (formData: User) => {
   const { _id: id, address, name, phone, isActive, isAdmin } = formData
