+++ src/app/dashboard/users/usersTableBody.tsx
@@ -4,8 +4,7 @@ import { Fragment, useEffect, useState } from 'react'
 
 import Image from 'next/image'
 
-import { User } from '@/app/types'
-
+import { User } from '@/types'
 import DeleteForm from '@/app/dashboard/users/delete-form'
 import ViewButton from '@/app/ui/dashboard/view-button/view-button'
 import { convertParseISODate } from '@/app/utils/convertParseISODate'
