+++ src/app/services/api-products.ts
@@ -1,5 +1,5 @@
-import { Product } from '../types'
-import { api } from './api'
+import { Product } from '@/types'
+import { api } from '@/app/services/api'
 
 export async function getProduct(id: string) {
   const { data } = await api.get(`/products/${id}`)
